Split out Visualization to separate MonoBehaviour
Resumption of development after break
