+++ AStar/Assets/Scripts/GameMap/MapTiles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40d0b2fdbc7d6f240886cac5c8bc0910
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
