+++ AStar/Assets/Scripts/GameMap/MapTiles/PathVisualizer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33e74d8c02b161a49b159cde10834a74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
